feat: make vitest tests pass
